fix(common): envFrom incorrectly generated configMapRef (#56)
We had a problem where the deployment template would not correctly
generate the configmap and secret envFrom references of the containers.
Everyone kept using the workaround of overriding the whole container
block of their chart.

From now on consumers of a chart should not have to override the
containers section of the chart when using fullnameOverride, the
generated envFrom references should be correct.

This pull request introduces updates to the `common` Helm chart,
including enhancements to the default handling of `envFrom` in container
configurations, new test cases, and additional sample values. The
changes improve flexibility in configuration and ensure proper testing
coverage for various scenarios.

### Updates to `common` Helm chart:

* **Version bump**: Updated the `common` chart version from `0.3.2` to
`0.3.3` in `Chart.yaml` and dependent charts (`common-unittests`) to
reflect the new changes.
[[1]](diffhunk://#diff-823513bda22d06799be5f9b3d83683efa999f4263575388d02303504811b396aL6-R6)
[[2]](diffhunk://#diff-bb7478441b6f39e4b8cb572a5c422e64b2c657efcac9ed99243e5f891f33942cL12-R12)

* **Default `envFrom` behavior**: Modified `_container.yaml` to provide
default `envFrom` values (configMapRef and secretRef) when not
explicitly defined in the container configuration.