fix(common): container.image object can now atomically be modified in values (#68)
In the previous iteration, when modifying the container.image object,
the template code did not consider all the cases.

For example, if you configured the name, but not the tag then the
versionOverride would not work.

This PR aims to fix most of these issues and unittests were added to
prevent regression in the future.

(autogenerated description)
This pull request introduces several updates to the `common` Helm chart
and its associated files. The changes include enhancements to the image
configuration logic, updates to the chart version, and new unit tests to
validate the image specifications. Additionally, the `index.yaml` file
is updated with new metadata for various chart versions.

### Helm Chart Updates:

* **Version Increment**: Updated the `common` Helm chart version from
`0.3.3` to `0.3.4` in `Chart.yaml` and other dependent charts
(`charts/tests/common-unittests/Chart.yaml`).
[[1]](diffhunk://#diff-823513bda22d06799be5f9b3d83683efa999f4263575388d02303504811b396aL6-R6)
[[2]](diffhunk://#diff-bb7478441b6f39e4b8cb572a5c422e64b2c657efcac9ed99243e5f891f33942cL12-R12)

### Image Configuration Enhancements:

* **Improved Image Handling**: Enhanced the `_container.yaml` template
to set default values for image attributes (`repository`, `name`, `tag`)
and added validation to prevent simultaneous use of `digest` and `tag`.
Defaults include `ghcr.io/interledger` for the repository and `latest`
for the tag unless overridden.

### Unit Tests:

* **Image Specification Tests**: Added a comprehensive suite of tests in
`image_specifications_test.yaml` to validate various scenarios for image
configuration, including default values, overrides, and digest usage.