feat(common): imageOverride mechanism at top level (#78)
This pull request introduces enhancements to the Helm chart for
Kubernetes resources, primarily focusing on adding support for image
overrides and updating dependencies. The most significant changes
include adding logic for image override configuration, updating the
chart version, and introducing a unit test to validate the new
functionality.

### Helm chart enhancements:

*
[`charts/common/templates/_container.yaml`](diffhunk://#diff-de315442c758d69ec1e19504903ecdbadb9c187c9f4bb86e92656a1d263858c5R12-R20):
Added logic to support overriding container image properties
(`repository`, `name`, `tag`, and `digest`) via `imageOverride` values.
Also ensured a default container name is set if not provided.

*
[`charts/tests/common-unittests/tests/image_override_test.yaml`](diffhunk://#diff-5c0ec171094fbdddc517b4e2045ce8866579e15b1d3398fcbe65d36199ce0b31R1-R14):
Introduced a new unit test to verify that the `imageOverride`
configuration is correctly applied to the container image.

### Dependency updates:

*
[`charts/common/Chart.yaml`](diffhunk://#diff-823513bda22d06799be5f9b3d83683efa999f4263575388d02303504811b396aL6-R6):
Updated the Helm chart version from `0.3.4` to `0.4.0` to reflect the
new features.

*
[`charts/tests/common-unittests/Chart.yaml`](diffhunk://#diff-bb7478441b6f39e4b8cb572a5c422e64b2c657efcac9ed99243e5f891f33942cL12-R12):
Updated the dependency version of the `common` chart to `0.4.0` to match
the updated chart version.