feat(ilf-wallet-backend): introduction of PDB and tolopogySpreadConstraints (#118)
To improve general availability of the service we introduce a minimum
replica count of 2, and also configure a PDB and topologySkew
contstraints to ensure that at least 1 other pod is available on an
alternative node.

The introduction of the `topologySpreadConstraints` should instruct
kubernetes to make sure that the pods are spread across nodes. This will
make us more resilient to involuntary node evictions.

At this point in time I decided to shy away from introducing horizontal
auto scaling because we are not confident of the ability of the
application to handle it. The purpose of the replication at the moment
is simply to improve up-time.

It is also worth noting that I resolved an issue in the charts there the
"app.kubernetes.io/instance" annotation was incorrectly set for the
backend server and worker deployments. Not sure what the effect of this
would have been.

Below are some snippets showing what the rendered templates will look
like now
```yaml
---
# Source: ilf-wallet-backend/templates/pdb.server.yaml
apiVersion: policy/v1
kind: PodDisruptionBudget
metadata:
  name: wallet-backend-ilf-wallet-backend-server
  labels:
    app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-server
    app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
spec:
  selector:
    matchLabels:
      app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-server
      app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
  minAvailable: 1
---
# Source: ilf-wallet-backend/templates/pdb.worker.yaml
apiVersion: policy/v1
kind: PodDisruptionBudget
metadata:
  name: wallet-backend-ilf-wallet-backend-worker
  labels:
    app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-worker
    app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
spec:
  selector:
    matchLabels:
      app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-worker
      app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
  minAvailable: 1
```

For the deployments
```yaml
# Source: ilf-wallet-backend/templates/deployment.server.yaml
apiVersion: apps/v1
kind: Deployment
metadata:
  labels:
    app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
    app.kubernetes.io/managed-by: Helm
    app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-server
    app.kubernetes.io/version: v1.2.8
    helm.sh/chart: ilf-wallet-backend-1.2.0
  name: wallet-backend-ilf-wallet-backend-server
spec:
  replicas: 2
  selector:
    matchLabels:
      app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
      app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-server
  template:
    metadata:
      labels:
        app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
        app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-server
    spec:
       ...(unchanged)...
      topologySpreadConstraints:
      - labelSelector:
          matchLabels:
            app.kubernetes.io/instance: wallet-backend-ilf-wallet-backend
            app.kubernetes.io/name: wallet-backend-ilf-wallet-backend-server
        maxSkew: 1
        topologyKey: kubernetes.io/hostname
        whenUnsatisfiable: DoNotSchedule