feat(common): support for segregated deployments and services (#166)
In previous versions of the common library, a consumer were forced to
always override the full container definition for a given deployment
whenever they wanted to override changes. This causes a lot of fragility
and duplication in real world deployments.

This PR introduces support for segregating deployments into separate
sections of the value files. The common framework will assume that the
main body of the deployment section also contains the parameters for the
pod itself. Additional containers can be defined using the `sidecars`
parameter.